requirements.txt
asgiref
beautifulsoup4~=4.9.3
certifi==2020.12.5
Django~=3.2
django-cors-middleware==1.5.0
pytz
regex~=2021.4.4
soupsieve
sqlparse
typing-extensions
openai~=0.6.3
pytrends~=4.7.3
Pillow~=8.2.0